driver: ioctl: remove force closing of handles
This driver has never actually made successful use of it, because we've
been wrongly matching against the FunctionalDeviceObject instead of
Stack->FileObject->DeviceObject. Yet, things seem to have worked fine
enough because of smart notification to the logger thread.

Furthermore, SwDevice calls halt immediately, because it constitutes a
surprise removal, which means we don't really even have time for the
fuse to go off.

And finally, dereferencing Table[i]->Object can race with the
destruction of that object, which is a UaF.

Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>